Move function to helper.py with other l10n functions
